Match MaximumCapsuleSize to UEFI spec
Per section 8.5.3.4 of the latest UEFI spec [1],
`QueryCapsuleCapabilities` uses a uint64 for
MaximumCapsuleSize, not a usize.

[1]:https://uefi.org/sites/default/files/resources/UEFI_Spec_2_10_Aug29.pdf

+++ uefi-raw/src/table/runtime.rs
@@ -66,7 +66,7 @@ pub struct RuntimeServices {
     pub query_capsule_capabilities: unsafe extern "efiapi" fn(
         capsule_header_array: *const *const CapsuleHeader,
         capsule_count: usize,
-        maximum_capsule_size: *mut usize,
+        maximum_capsule_size: *mut u64,
         reset_type: *mut ResetType,
     ) -> Status,
